feat: add count to project dashboard
This change removes the example page and adds a component to display the
number of Kamelets running in the current project on the project
overview page in the okd/openshift console.  The count is displayed as a
link like other resources in the Inventory card and links to a list view
of available kamelets in the current project.