feat: add some quick POC hacks
This change adds a replacement create page for Kamelets with some POC
screenshots to illustrate a page flow.  It also bumps the SDK dependency
to the latest to get access to the YAML resource editor and uses this in
the create page to provide a way to create a Kamelet from text still.