feat: embed build info
This change updates the webpack config to include build info, along with
logging in the app so it's easy to tell what version is deployed.  A
build script is now included to automate pushing to the cluster's image
registry.  This change also switches back to using the SDK's
ListPageCreate button for the empty state.